cmake: Make DBUS_SESSION_SOCKET_DIR in CMakeCache.txt consistent with config.h
It's unexpected to have DBUS_SESSION_SOCKET_DIR take different values in
CMakeCache.txt, which lists the variables that are available to be set
by the caller, and in config.h, which makes their final values available
to the C code. If DBUS_SESSION_SOCKET_DIR is empty and we are running
on Unix, set it to its dynamically-chosen fallback before storing it in
the cache.

[smcv: Add commit message]
Co-authored-by: Simon McVittie <smcv@collabora.com>